python/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-road"
version = "0.1.0"
description = "Low-frequency quant MVP aligned with docs/req.md"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "akshare>=1.16.0",
  "backtrader>=1.9.78.123",
  "pandas>=2.2.0",
  "psycopg2-binary>=2.9.9",
  "redis>=5.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
